fix memory issues

+++ api/matlab/loc_destroy.m
@@ -10,7 +10,7 @@ op_code = loc_opcodes( 'destroy' );
 if ( nargin < 1 || isempty(loc) )
   loc_api( op_code );
 else
-  loc_api( op_code, loc );
+  loc_api( op_code, uint32(loc) );
 end
 
 end
